Relocate ref_deleted webhook processor out of dashboard_templates
Pure move: ``RefDeletedWebhookProcessor`` and ``RefDeletedResult``
shipped under ``services/dashboard_templates/`` because GitHub webhook
processing started there, not because the rule has anything to do with
dashboard templates. The class lives in the base ``services/``
hierarchy now so PRD #346's webhook unpublish/dashboard feature (#389)
can grow it in place rather than landing in a misleading directory.

Key decisions:
- Pure relocation, no behaviour change. ``git mv`` preserves the file
  history and the test suite is unchanged beyond its import path. No
  re-export shim left behind in ``dashboard_templates/__init__.py``:
  the only in-tree importers are ``factory.py`` and
  ``handlers/webhooks/github.py``, both updated to point at the new
  module, so a deprecation alias would only add ceremony.
- The internal ``_coerce_int`` docstring still cross-refs
  ``services.dashboard_templates.push_processor._coerce_int``. That
  target genuinely lives in dashboard_templates today (the push
  processor is still bound to the dashboard-template binding contract),
  so the cross-ref is correct as-written.
- Sphinx cross-refs in ``storage/edition_store.py``,
  ``storage/project_store.py``, and the matching docstring in
  ``tests/storage/edition_store_test.py`` updated to the new dotted
  path so ``intersphinx``/IDE jumps don't 404.

Next-iteration notes:
- The follow-up webhook-unpublish/dashboard work from PRD #346 (#389)
  now lands directly in ``services/ref_deleted_processor.py``; no
  pre-emptive scaffolding here.

Closes #388
