+++ src/docverse/services/ref_deleted_processor.py
